Updating the rollback tool so that it supports rolling back by workflow rerun
When the selected project has no rollback pipeline definition configured, we are going to ask the user if they want to rerun a workflow. The rest of the flow will be the same as the normal rollback flow, but, at the end, we will trigger a workflow rerun based on the workflow id associated with the selected version.